refactored intensity processing script
